[VAS] : ITEM 9774 : refactoring archive service : backend and front end archive search
