+++ ui/ui-frontend/projects/archive-search/src/app/archive/management-rules/archive-unit-rules/add-management-rules/add-management-rules.component.ts
@@ -349,6 +349,7 @@ export class AddManagementRulesComponent implements OnInit, OnDestroy {
   }
 
   submit() {
+    this.resultNumberToShow = this.translateService.instant('ARCHIVE_SEARCH.MORE_THAN_THRESHOLD');
     this.isDisabled = true;
     this.showText = true;
     this.isLoading = !this.isLoading;
