+++ ui/ui-frontend/projects/archive-search/src/app/archive/management-rules/archive-unit-rules/delete-unit-rules/delete-unit-rules.component.html
@@ -55,12 +55,12 @@
 </ng-template>
 
 <ng-template #okTotalFound>
-  <div *ngIf="showText && itemsWithSameRule > 0" class="row message-text-ok">
+  <div *ngIf="showText && itemsWithSameRule !== 0" class="row message-text-ok">
     <div class="col-1"><i class="material-icons">check_circle</i></div>
     <div class="col-11">{{ itemsWithSameRule }} {{ 'RULES.APRAISAL_RULES.UA_TO_UPDATE' | translate }}</div>
   </div>
 
-  <div *ngIf="showText && itemsToNotUpdate > 0" class="row message-text-ko">
+  <div *ngIf="showText && itemsToNotUpdate !== 0" class="row message-text-ko">
     <div class="col-1"><i class="material-icons">error</i></div>
     <div class="col-11">{{ itemsToNotUpdate }} {{ 'RULES.APRAISAL_RULES.UA_TO_IGNORE' | translate }}</div>
   </div>
