+++ ui/ui-frontend/projects/archive-search/src/app/archive/management-rules/archive-unit-rules/delete-unit-rules/delete-unit-rules.component.ts
@@ -185,6 +185,7 @@ export class DeleteUnitRulesComponent implements OnInit, OnDestroy {
     this.disabledControl = true;
     this.showText = true;
     this.isLoading = !this.isLoading;
+    this.resultNumberToShow = this.translateService.instant('ARCHIVE_SEARCH.MORE_THAN_THRESHOLD');
 
     const rule: RuleAction = {
       rule: this.ruleDetailsForm.get('rule').value,
